Added a function to call pubsub when a package is requested
This custom code will configure a message to be sent to a google pub/sub topic when a package is first requested. This message will only be sent once when the package is downloaded from PyPi to the pulp cache and will never be sent again.

To make sure this works, you need to have the `GOOGLE_APPLICATION_CREDENTIALS` setting set for the environment where pulp is running. You will also need to configure the Install this plugin through pip, reboot the `GOOGLE_PUBSUB_PROJECT_ID` and `GOOGLE_PUBSUB_TOPIC_ID` variables from the django settings, this should be set in the pulp settings file. Reboot the content service and try downloading any package, it should send a message to your topic.

An ansible playbook and some instructions will set all this up through goval, so this should never be used manually.